sdk: reorganise
