+++ src/ui/results/index.js
@@ -1,7 +1,10 @@
-export { default as Results } from "./Results";
-export { default as Result } from "./Result";
-export { default as Title } from "./Title";
 export { default as Description } from "./Description";
-export { default as Url } from "./Url";
-export { default as TokenLink } from "./TokenLink";
 export { default as Image } from "./Image";
+export { default as Paginator } from "./Paginator";
+export { default as Response } from "./Response";
+export { default as Result } from "./Result";
+export { default as Results } from "./Results";
+export { default as Summary } from "./Summary";
+export { default as Title } from "./Title";
+export { default as TokenLink } from "./TokenLink";
+export { default as Url } from "./Url";
