+++ src/ui/text/Input.js
@@ -1,0 +1,93 @@
+import React from "react";
+import { findDOMNode } from "react-dom";
+
+import { changeEvent } from "../../controllers/values";
+import { resultsEvent } from "../../controllers/pipeline";
+
+const RETURN_KEYCODE = 13;
+
+class Input extends React.Component {
+  constructor(props) {
+    super(props);
+    this.setText = this.setText.bind(this);
+    this.valuesUpdated = this.valuesUpdated.bind(this);
+    this.getState = this.getState.bind(this);
+    this.state = {
+      ...this.getState(props.values, props.pipeline, props.qParam),
+      qParam: props.qParam,
+      qOverrideParam: props.qOverrideParam
+    };
+  }
+
+  getState(values, pipeline, qParam) {
+    const text = values.get()[qParam] || "";
+    return { text };
+  }
+
+  componentDidMount() {
+    if (this.props.focus) {
+      findDOMNode(this.refs.searchInput).focus();
+    }
+
+    this.removeValuesListener = this.props.values.listen(
+      changeEvent,
+      this.valuesUpdated
+    );
+  }
+
+  componentWillUnmount() {
+    this.removeValuesListener();
+  }
+
+  valuesUpdated() {
+    this.setState(
+      this.getState(this.props.values, this.props.pipeline, this.state.qParam)
+    );
+  }
+
+  setText(text) {
+    const textValues = {
+      [this.state.qParam]: text,
+      [this.state.qOverrideParam]: "true"
+    };
+    this.props.values.set(textValues);
+    if (textValues[this.state.qParam]) {
+      this.props.pipeline.search();
+    }
+  }
+
+  render() {
+    const { text } = this.state;
+    const { placeHolder, className, instant } = this.props;
+
+    return (
+      <input
+        type="text"
+        ref="searchInput"
+        className={className}
+        placeholder={placeHolder}
+        value={text}
+        onChange={e => {
+          if (instant) {
+            this.setText(e.target.value);
+          }
+        }}
+        onKeyDown={e => {
+          if (e.keyCode === RETURN_KEYCODE) {
+            e.preventDefault();
+            this.setText(text);
+          }
+        }}
+      />
+    );
+  }
+}
+
+Input.defaultProps = {
+  qParam: "q",
+  qOverrideParam: "q.override",
+  placeHolder: "Type to search",
+  instant: true
+};
+
+export default Input;
